Prefix variables names
This PR is adding a `REGISTRY_` prefix to the `HOSTNAME`, `USERNAME`, and `PASSWORD` variables.

This is needed to prevent the plugin from overriding the `HOSTNAME` and `USERNAME` system environment variables.

Also add support for cosign, podman & skopeo cli tools and fix the docker config file path in the log message